+++ src/operators/tensor/implementations/tensor_fp8x23wide.cairo
@@ -300,7 +300,7 @@ impl FP8x23WTensor of TensorTrait<FP8x23W> {
     }
 
     fn gather(
-        self: @Tensor<FP8x23W>, indices: Tensor<usize>, axis: Option<usize>
+        self: @Tensor<FP8x23W>, indices: Tensor<i32>, axis: Option<i32>
     ) -> Tensor<FP8x23W> {
         math::gather::gather(self, indices, axis)
     }
