+++ src/operators/tensor/implementations/tensor_i32.cairo
@@ -340,7 +340,7 @@ impl I32Tensor of TensorTrait<i32> {
         core_tensor::slice::<i32>(self, starts, ends, axes, steps)
     }
 
-    fn gather(self: @Tensor<i32>, indices: Tensor<usize>, axis: Option<usize>) -> Tensor<i32> {
+    fn gather(self: @Tensor<i32>, indices: Tensor<i32>, axis: Option<i32>) -> Tensor<i32> {
         math::gather::gather(self, indices, axis)
     }
